Adds CEO button and makes group of revenue bonuses
Changed the order so that the garden button goes with the cfo bonus
since these both apply to the revenue but not the unit price. I put them
below the demand part too since they are "add-ons" to the demand
calculatoin, again making the logic of calculation slightly more
implicitly clear to the user.