Refactors to pass toggle button colors
The ToggleButton component now includes an optional string argument for
the color instead of the flag for whether it should be pink. The default
is set as the existing default blue.

I also created an object to provide rgb colors by name that can be
imported when colors are needed, eg. ToggleButtonColors.purple.

The CFO button is now purple.