Puts color palette in separate file
Color palette has its own file and there is now a FoodChain
MagnatePalette.

This is labelled wth the type of the component targetted rather than the
color, e.g. cfo rather than purple.

I also added defaults that are now used in the rest of the JS, so no rgb
values are to be found except in the color palette and the CSS - this
will make it easier to change colours later on.