Revenue bonuses (#11)
* Fixes issue #1: '-1$ Sale!' button now pale pink

Added optional parameter to the toggle button component - boolean to say
whether component is a permanent sale button or not. Then fixed the
toggle button component to be pink if it's a permanent sale button and
blue otherwise.

* Adds CEO button and makes group of revenue bonuses

Changed the order so that the garden button goes with the cfo bonus
since these both apply to the revenue but not the unit price. I put them
below the demand part too since they are "add-ons" to the demand
calculatoin, again making the logic of calculation slightly more
implicitly clear to the user.

* Refactors to pass toggle button colors

The ToggleButton component now includes an optional string argument for
the color instead of the flag for whether it should be pink. The default
is set as the existing default blue.

I also created an object to provide rgb colors by name that can be
imported when colors are needed, eg. ToggleButtonColors.purple.

The CFO button is now purple.

* Puts color palette in separate file

Color palette has its own file and there is now a FoodChain
MagnatePalette.

This is labelled wth the type of the component targetted rather than the
color, e.g. cfo rather than purple.

I also added defaults that are now used in the rest of the JS, so no rgb
values are to be found except in the color palette and the CSS - this
will make it easier to change colours later on.