Adds horizontal option for number buttons
Added via a flag that defaults to false - I already set it to
true for the demand buttons but the unit price ones remain
vertical.

This saves space in the case of the demand icons - there still
needs to be formatting changes e.g. the fact that the buttons
demand buttons aren't quite aligned due to differently sized svgs
but that's for a different commit.