:sparkles: Adds waitress and waitress bonus (#21)
* Adds waitress and waitress bonus

Formatting not done well yet but the buttons are
logic are there.

* Removes package-lock.json

Accidentally kept after pulling from the upstream repo post change to
yarn - now deleted to avoid npm conflict.

* Makes first waitress a milestone

Rather than it being a separate boolean, it's just part of the
milestones object. The result is the same but it's more intuitive and
avoids the need for an entirely different object to be created.